+++ ntex/src/web/ws.rs
@@ -1,7 +1,7 @@
 use std::error::Error as StdError;
 
 use bytes::Bytes;
-use futures::{Stream, TryStreamExt};
+use futures::{Sink, Stream, TryStreamExt};
 
 pub use crate::ws::{CloseCode, CloseReason, Frame, Message};
 
@@ -17,9 +17,9 @@ pub type WebSocketsSink =
 
 /// Do websocket handshake and start websockets service.
 pub async fn start<T, F, S, Err>(
-    factory: F,
     req: HttpRequest,
     payload: S,
+    factory: F,
 ) -> Result<HttpResponse, Err>
 where
     T: ServiceFactory<
@@ -35,17 +35,40 @@ where
     Err: From<T::InitError>,
     Err: From<HandshakeError>,
 {
+    let (tx, rx) = mpsc::channel();
+
+    start_with(req, payload, tx, rx, factory).await
+}
+
+/// Do websocket handshake and start websockets service.
+pub async fn start_with<T, F, S, Err, Tx, Rx>(
+    req: HttpRequest,
+    payload: S,
+    tx: Tx,
+    rx: Rx,
+    factory: F,
+) -> Result<HttpResponse, Err>
+where
+    T: ServiceFactory<
+        Config = ws::StreamEncoder<Tx>,
+        Request = Frame,
+        Response = Option<Message>,
+    >,
+    T::Error: StdError + 'static,
+    T::InitError: 'static,
+    T::Service: 'static,
+    F: IntoServiceFactory<T>,
+    S: Stream<Item = Result<Bytes, PayloadError>> + Unpin + 'static,
+    Err: From<T::InitError>,
+    Err: From<HandshakeError>,
+    Tx: Sink<Result<Bytes, Box<dyn StdError>>> + Clone + Unpin + 'static,
+    Tx::Error: StdError,
+    Rx: Stream<Item = Result<Bytes, Box<dyn StdError>>> + Unpin + 'static,
+{
     // ws handshake
     let mut res = handshake(req.head())?;
 
-    let payload = payload.map_err(|e| {
-        let e: Box<dyn StdError> = Box::new(e);
-        e
-    });
-
     // response body stream
-    let (tx, rx): (_, mpsc::Receiver<Result<Bytes, Box<dyn StdError>>>) =
-        mpsc::channel();
     let sink = ws::StreamEncoder::new(tx);
 
     // create ws service